chore: rename til file-input
Matcher Figma, og komponenten har ikke ansvar for selve opplastingen,
kun input.
