+++ packages/file-input-react/documentation/iconBytes.ts
