+++ packages/file-input-react/integration/__image_snapshots__/FileUploader renders correctly #0.png
