+++ packages/file-input-react/integration/__image_snapshots__/FileUploader renders correctly in high contrast #0.png
