+++ packages/file-input-react/src/internal/Dropzone.tsx
@@ -1,15 +1,15 @@
 import { WithChildren } from "@fremtind/jkl-core";
 import cn from "classnames";
 import React, { forwardRef, useState } from "react";
-import { FileUploadState } from "../types";
-import { useFileUploaderContext } from "./fileUploaderContext";
+import { FileState } from "../types";
+import { useFileInputContext } from "./fileInputContext";
 import { validateFile } from "./validateFile";
 
 interface DropzoneProps extends WithChildren {}
 
 export const Dropzone = forwardRef<HTMLDivElement, DropzoneProps>((props, ref) => {
     const { children, ...rest } = props;
-    const { maxSizeBytes, accept, onChange } = useFileUploaderContext();
+    const { maxSizeBytes, accept, onChange } = useFileInputContext();
 
     const [onDragClassName, setOnDragClassName] = useState<string>("");
 
@@ -17,14 +17,14 @@ export const Dropzone = forwardRef<HTMLDivElement, DropzoneProps>((props, ref) =
         <div
             {...rest}
             ref={ref}
-            className={cn("jkl-file-uploader__dropzone", onDragClassName)}
+            className={cn("jkl-file-input__dropzone", onDragClassName)}
             onDragEnter={(e) => {
-                setOnDragClassName("jkl-file-uploader__dropzone--enter");
+                setOnDragClassName("jkl-file-input__dropzone--enter");
                 e.preventDefault();
             }}
             onDragOver={(e) => {
                 /* Prevent browser from opening file in a new tab */
-                setOnDragClassName("jkl-file-uploader__dropzone--enter");
+                setOnDragClassName("jkl-file-input__dropzone--enter");
                 e.preventDefault();
             }}
             onDrop={(e) => {
@@ -34,7 +34,7 @@ export const Dropzone = forwardRef<HTMLDivElement, DropzoneProps>((props, ref) =
                 if (e.dataTransfer.files) {
                     onChange(
                         e,
-                        [...e.dataTransfer.files].map<FileUploadState>((file) => ({
+                        [...e.dataTransfer.files].map<FileState>((file) => ({
                             file,
                             isUploading: false,
                             validation: validateFile(file, accept, maxSizeBytes),
