+++ packages/file-input-react/src/types.ts
@@ -1,0 +1,10 @@
+export interface FileValidation {
+    type: "TOO_BIG" | "WRONG_FORMAT";
+    message: string;
+}
+
+export interface FileState {
+    file: File;
+    validation?: FileValidation;
+    isUploading: boolean;
+}
