+++ packages/file-input-react/tsconfig-for-declarations.json
