+++ packages/file-input/README.md
@@ -1,0 +1,20 @@
+# [`@fremtind/jkl-file-input`](https://jokul.fremtind.no/komponenter/file-input)
+
+Se portalen for [bruk og prinsipper](https://jokul.fremtind.no/komponenter/file-input).
+
+## Installasjon
+
+**Tips:** om du bruker [React-pakken](../file-input-react/) trenger du ikke installere denne pakken direkte.
+
+1. `npm i @fremtind/jkl-file-input`.
+2. Importér stil-pakken i prosjektet ditt.
+
+```js
+// Importer stilark via JavaScript med CSS-loader.
+import "@fremtind/jkl-file-input/file-input.min.css";
+```
+
+```scss
+// Eller importer stilark via SCSS.
+@use "@fremtind/jkl-file-input/file-input";
+```
