+++ packages/file-input/_index.scss
@@ -1,0 +1,1 @@
+@forward "file-input";
