+++ packages/file-input/package.json
@@ -1,11 +1,11 @@
 {
-    "name": "@fremtind/jkl-file-uploader",
+    "name": "@fremtind/jkl-file-input",
     "version": "1.0.0-alpha.0",
     "publishConfig": {
         "access": "public"
     },
-    "description": "Jøkul style for file-uploader",
-    "homepage": "https://github.com/fremtind/jokul/tree/main/packages/file-uploader",
+    "description": "Jøkul style for file-input",
+    "homepage": "https://github.com/fremtind/jokul/tree/main/packages/file-input",
     "keywords": [
         "jøkul",
         "fremtind"
@@ -20,7 +20,7 @@
         "build": "gulp build",
         "build:watch": "gulp build:watch",
         "test": "echo \"Error: run tests from root\" && exit 1",
-        "dev": "echo \"Error: run dev from file-uploader-react\" && exit 1"
+        "dev": "echo \"Error: run dev from file-input-react\" && exit 1"
     },
     "dependencies": {
         "@fremtind/jkl-core": "^12.0.2"
