--- packages/file-uploader-react/src/index.ts
@@ -1,5 +1,0 @@
-export type { FileProps } from "./File";
-export { File } from "./File";
-export type { FileUploaderProps } from "./FileUploader";
-export { FileUploader } from "./FileUploader";
-export type { FileUploadState, FileUploadValidation } from "./types";
