--- packages/file-uploader-react/src/types.ts
@@ -1,10 +1,0 @@
-export interface FileUploadValidation {
-    type: "TOO_BIG" | "WRONG_FORMAT";
-    message: string;
-}
-
-export interface FileUploadState {
-    file: File;
-    validation?: FileUploadValidation;
-    isUploading: boolean;
-}
