--- packages/file-uploader/README.md
@@ -1,20 +1,0 @@
-# [`@fremtind/jkl-file-uploader`](https://jokul.fremtind.no/komponenter/file-uploader)
-
-Se portalen for [bruk og prinsipper](https://jokul.fremtind.no/komponenter/file-uploader).
-
-## Installasjon
-
-**Tips:** om du bruker [React-pakken](../file-uploader-react/) trenger du ikke installere denne pakken direkte.
-
-1. `npm i @fremtind/jkl-file-uploader`.
-2. Importér stil-pakken i prosjektet ditt.
-
-```js
-// Importer stilark via JavaScript med CSS-loader.
-import "@fremtind/jkl-file-uploader/file-uploader.min.css";
-```
-
-```scss
-// Eller importer stilark via SCSS.
-@use "@fremtind/jkl-file-uploader/file-uploader";
-```
