--- packages/file-uploader/_index.scss
@@ -1,1 +1,0 @@
-@forward "file-uploader";
