--- packages/file-uploader/file-uploader.scss
@@ -1,172 +1,0 @@
-@use "~@fremtind/jkl-core/jkl";
-@use "~@fremtind/jkl-core/jkl/colors";
-
-@include jkl.light-mode-variables {
-    --jkl-file-uploader-color: var(--jkl-color, #{jkl.$color-granitt});
-    --jkl-file-uploader-drop-border-color: #{jkl.$color-fjellgra};
-    --jkl-file-uploader-background-color-active: #{jkl.$color-varde};
-    --jkl-file-uploader-border-color: #{jkl.$color-dis};
-    --jkl-file-uploader-support-color: #{jkl.$color-stein};
-    --jkl-file-uploader-thumbnail-background-color: #{jkl.$color-sand};
-}
-
-@include jkl.dark-mode-variables {
-    --jkl-file-uploader-color: var(--jkl-color, #{jkl.$color-snohvit});
-    --jkl-file-uploader-drop-border-color: #{jkl.$color-varde};
-    --jkl-file-uploader-background-color-active: #{jkl.$color-stein};
-    --jkl-file-uploader-border-color: #{jkl.$color-dis};
-    --jkl-file-uploader-support-color: #{jkl.$color-svaberg};
-    --jkl-file-uploader-thumbnail-background-color: #{jkl.$color-stein};
-}
-
-@include jkl.comfortable-density-variables {
-    @include jkl.declare-font-variables("jkl-file-uploader-drag", "heading-2");
-
-    --jkl-file-uploader-input-padding: #{jkl.$spacing-40};
-    --jkl-file-uploader-input-gap: #{jkl.$spacing-24};
-}
-
-@include jkl.compact-density-variables {
-    @include jkl.declare-font-variables("jkl-file-uploader-drag", "body");
-
-    --jkl-file-uploader-input-padding: #{jkl.$spacing-16};
-    --jkl-file-uploader-input-gap: #{jkl.$spacing-8};
-}
-
-.jkl-file-uploader-input {
-    border: var(--jkl-file-uploader-drop-border-color) 1px dashed;
-    border-radius: jkl.rem(4px);
-    padding: var(--jkl-file-uploader-input-padding);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-flow: column nowrap;
-    gap: var(--jkl-file-uploader-input-gap);
-    @include jkl.motion("standard", "productive");
-    transition-property: background-color;
-    background-color: transparent;
-
-    &--enter {
-        background-color: var(--jkl-file-uploader-background-color-active);
-        border-style: solid;
-    }
-
-    &__drag-text {
-        @include jkl.use-font-variables("jkl-file-uploader-drag");
-        @include jkl.small-device {
-            display: none;
-        }
-    }
-
-    &__max-size-text {
-        color: var(--jkl-file-uploader-color);
-    }
-
-    html:not([data-mousenavigation]):not([data-touchnavigation]) &:focus-within {
-        .jkl-button {
-            transform: scale(1.05);
-
-            &::after {
-                box-shadow: 0 0 0 jkl.rem(2px) var(--jkl-button-focus-color);
-            }
-
-            &:active {
-                transform: scale(1);
-            }
-        }
-    }
-}
-
-.jkl-file-uploader-previews {
-    list-style: none;
-    margin: 0;
-    padding: 0;
-
-    &__item {
-        margin: 0;
-        padding: 0;
-    }
-}
-
-@include jkl.comfortable-density-variables {
-    --jkl-file-uploader-preview-padding: #{jkl.$spacing-16} #{jkl.$spacing-12};
-    --jkl-file-uploader-preview-thumbnail-size: #{jkl.rem(64px)};
-    --jkl-file-uploader-preview-gap: #{jkl.$spacing-16};
-}
-
-@include jkl.compact-density-variables {
-    --jkl-file-uploader-preview-padding: #{jkl.$spacing-12} #{jkl.$spacing-8};
-    --jkl-file-uploader-preview-thumbnail-size: #{jkl.rem(44px)};
-    --jkl-file-uploader-preview-gap: #{jkl.$spacing-8};
-}
-
-.jkl-file-uploader-preview {
-    @include jkl.reset-outline;
-    @include jkl.text-style("small");
-    color: var(--jkl-file-uploader-color);
-    border-bottom: 1px var(--jkl-file-uploader-border-color) solid;
-    padding: var(--jkl-file-uploader-preview-padding);
-
-    &[href] {
-        cursor: pointer;
-        text-decoration: none;
-        display: block;
-
-        &:hover .jkl-file-uploader-preview__file-size {
-            @include jkl.no-grow-bold;
-        }
-
-        @include jkl.keyboard-navigation {
-            &:focus {
-                outline: jkl.rem(2px) solid var(--jkl-file-uploader-color);
-            }
-        }
-
-        @include jkl.forced-colors-mode {
-            outline: revert;
-            border: jkl.rem(2px) solid LinkText;
-        }
-    }
-
-    &:last-of-type {
-        border-bottom: none;
-    }
-
-    &--invalid {
-        --jkl-label-error-color: #{jkl.$color-granitt}; // Vises på rød varslingsfarge
-        background-color: colors.varslingsfarge("feil");
-        color: jkl.$color-granitt;
-    }
-
-    &__info-wrapper {
-        display: flex;
-        flex-flow: nowrap row;
-        gap: var(--jkl-file-uploader-preview-gap);
-        align-items: center;
-    }
-
-    &__thumbnail {
-        width: var(--jkl-file-uploader-preview-thumbnail-size);
-        height: var(--jkl-file-uploader-preview-thumbnail-size);
-        flex: 0 0 var(--jkl-file-uploader-preview-thumbnail-size);
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        object-fit: contain;
-        border-radius: jkl.rem(4px);
-    }
-
-    &__thumbnail:not(img) {
-        background: var(--jkl-file-uploader-thumbnail-background-color);
-    }
-
-    &__file-info {
-        flex: 1 1 auto;
-        overflow-wrap: break-word;
-    }
-
-    &__file-name {
-        @include jkl.text-style("heading-5");
-        overflow-wrap: anywhere;
-    }
-}
