+++ packages/loader-react/documentation/Loader.mdx
@@ -61,7 +61,7 @@ Dersom handlingen vanligvis tar under et halvt sekund, men _kan_ ta lenger tid e
     <DoDontExample type="dont" description="Vis et kort glimt av en loader" content={<QuickLoaderExampleDont />} />
 </Grid>
 
-Gi brukeren tilbakemelding dersom handlingen tar lang tid (over fem sekunder). Om du har en målbar prosess, for eksempel [filopplasting](/komponenter/fileuploader), bør du gi brukeren tilbakemelding på hvor langt prosessen har kommet, og hvor langt de har igjen. Se [ProgressBar](/komponenter/progressbar).
+Gi brukeren tilbakemelding dersom handlingen tar lang tid (over fem sekunder). Om du har en målbar prosess, for eksempel [filopplasting](/komponenter/fileinput), bør du gi brukeren tilbakemelding på hvor langt prosessen har kommet, og hvor langt de har igjen. Se [ProgressBar](/komponenter/progressbar).
 
 ## Skeleton loader
